+++ src/main/java/com/good/food/core/entity/ClienteEntity.java
@@ -6,16 +6,16 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.RequiredArgsConstructor;
 
 @Entity
 @Data
 @Builder
-@RequiredArgsConstructor
 @AllArgsConstructor
+@Table(name = "cliente")
 public class ClienteEntity {
 
   @Id
