+++ src/main/java/com/good/food/core/entity/PedidoEntity.java
@@ -17,16 +17,16 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.RequiredArgsConstructor;
 
 @Entity
 @Data
 @Builder
-@RequiredArgsConstructor
 @AllArgsConstructor
+@Table(name = "pedido")
 public class PedidoEntity {
 
   @Id
