+++ src/main/java/com/good/food/core/entity/ProdutoEntity.java
@@ -9,16 +9,16 @@ import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
-import lombok.RequiredArgsConstructor;
 
 @Entity
 @Data
 @Builder
-@RequiredArgsConstructor
 @AllArgsConstructor
+@Table(name = "produto")
 public class ProdutoEntity {
 
   @Id
