+++ src/main/java/com/good/food/core/usecase/AdicionarAoPedido.java
@@ -1,34 +1,35 @@
-package com.good.food.usecase.pedido;
+package com.good.food.core.usecase;
 
-import java.util.List;
-import org.springframework.beans.factory.annotation.Autowired;
+import java.util.UUID;
 import org.springframework.stereotype.Component;
 import com.good.food.domain.Pedido;
 import com.good.food.domain.exceptions.NotFoundException;
 import com.good.food.gateways.PedidoDatabaseGateway;
+import com.good.food.gateways.http.request.PedidoRequest;
 import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class RemoverDoPedido {
+public class AdicionarAoPedido {
 
-  @Autowired
   private final PedidoDatabaseGateway pedidoDatabaseGateway;
-
-  @Autowired
   private final BuscarPedido buscarPedido;
+  private final CadastrarItemPedido cadastrarItemPedido;
 
-  public Pedido execute(final String uuidPedido, final List<String> itemsToBeRemoved) {
+  public Pedido execute(final String uuidPedido, final PedidoRequest pedidoRequest) {
     Pedido pedido;
 
     try {
-      pedido = buscarPedido.execute(uuidPedido);
+      pedido = buscarPedido.execute(UUID.fromString(uuidPedido));
     } catch (Exception e) {
       throw new NotFoundException("Pedido não existente.", e);
     }
 
-    itemsToBeRemoved.forEach(pedido::removerItem);
+    pedidoRequest.getItemPedidos().forEach(itemPedidoRequest -> {
+      cadastrarItemPedido.execute(pedido, itemPedidoRequest);
+    });
 
     return pedidoDatabaseGateway.save(pedido);
   }
+
 }
