+++ src/main/java/com/good/food/core/usecase/BuscarPedido.java
@@ -1,5 +1,6 @@
-package com.good.food.usecase.pedido;
+package com.good.food.core.usecase;
 
+import java.util.UUID;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import com.good.food.domain.Pedido;
@@ -8,15 +9,12 @@ import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class AvancarStatus {
+public class BuscarPedido {
 
   @Autowired
   private final PedidoDatabaseGateway pedidoDatabaseGateway;
 
-  public Pedido execute(String pedidoId) {
-    Pedido pedidoAtual = pedidoDatabaseGateway.findById(pedidoId);
-    //Avaliar
-//    pedidoAtual.setStatus(pedidoAtual.getStatus().next());
-    return pedidoDatabaseGateway.save(pedidoAtual);
+  public Pedido execute(UUID uuid) {
+    return pedidoDatabaseGateway.findById(uuid);
   }
 }
