+++ src/main/java/com/good/food/core/usecase/BuscarProdutoPorCategoria.java
@@ -1,4 +1,4 @@
-package com.good.food.usecase.produto;
+package com.good.food.core.usecase;
 
 import java.util.List;
 import java.util.stream.Collectors;
