+++ src/main/java/com/good/food/core/usecase/CadastrarPedido.java
@@ -1,10 +1,10 @@
-package com.good.food.usecase.pedido;
+package com.good.food.core.usecase;
 
 import org.springframework.stereotype.Component;
 import com.good.food.domain.Pedido;
 import com.good.food.gateways.PedidoDatabaseGateway;
 import com.good.food.gateways.http.request.PedidoRequest;
-import com.good.food.usecase.cliente.BuscarCliente;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 
 @Component
@@ -15,6 +15,7 @@ public class CadastrarPedido {
   private final BuscarCliente buscarCliente;
   private final CadastrarItemPedido cadastrarItemPedido;
 
+  @Transactional
   public Pedido execute(final PedidoRequest pedidoRequest) {
     final Pedido pedido = pedidoRequest.toDomain();
     pedido.setCliente(buscarCliente.execute(pedidoRequest.getClienteCPF()));
@@ -22,12 +23,14 @@ public class CadastrarPedido {
     // useCase futuro pra pagamento
     pedido.setStatus("RECEBIDO");
     
+    final Pedido pedidoSaved = pedidoDatabaseGateway.save(pedido);
+    
     pedidoRequest.getItemPedidos()
       .forEach(itemPedidoRequest -> {
-        pedido.addItem(cadastrarItemPedido.execute(pedido, itemPedidoRequest));
+        pedidoSaved.getItemPedido().add(cadastrarItemPedido.execute(pedidoSaved, itemPedidoRequest));
       });
 
-    return pedidoDatabaseGateway.save(pedido);
+    return pedidoSaved;
   }
 
 }
