+++ src/main/java/com/good/food/core/usecase/RemoverDoPedido.java
@@ -1,35 +1,35 @@
-package com.good.food.usecase.pedido;
+package com.good.food.core.usecase;
 
+import java.util.List;
+import java.util.UUID;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import com.good.food.core.entity.PedidoEntity;
 import com.good.food.domain.Pedido;
 import com.good.food.domain.exceptions.NotFoundException;
 import com.good.food.gateways.PedidoDatabaseGateway;
-import com.good.food.gateways.http.request.PedidoRequest;
 import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class AdicionarAoPedido {
+public class RemoverDoPedido {
 
+  @Autowired
   private final PedidoDatabaseGateway pedidoDatabaseGateway;
-  private final BuscarPedido buscarPedido;
-  private final CadastrarItemPedido cadastrarItemPedido;
 
-  public Pedido execute(final String uuidPedido, final PedidoRequest pedidoRequest) {
+  @Autowired
+  private final BuscarPedido buscarPedido;
+
+  public Pedido execute(final String uuidPedido, final List<String> itemsToBeRemoved) {
     Pedido pedido;
 
     try {
-      pedido = buscarPedido.execute(uuidPedido);
+      pedido = buscarPedido.execute(UUID.fromString(uuidPedido));
     } catch (Exception e) {
       throw new NotFoundException("Pedido não existente.", e);
     }
 
-    pedidoRequest.getItemPedidos().forEach(itemPedidoRequest -> {
-      cadastrarItemPedido.execute(pedido, itemPedidoRequest);
-    });
+//    itemsToBeRemoved.forEach(pedido::removerItem);
 
     return pedidoDatabaseGateway.save(pedido);
   }
-
 }
