+++ src/main/java/com/good/food/gateways/http/ProdutoController.java
@@ -22,12 +22,12 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import com.good.food.core.entity.EProdutoCategoria;
+import com.good.food.core.usecase.BuscarProdutoPorCategoria;
+import com.good.food.core.usecase.CadastrarProduto;
+import com.good.food.core.usecase.EditarProduto;
+import com.good.food.core.usecase.RemoverProduto;
 import com.good.food.gateways.http.request.ProdutoRequest;
 import com.good.food.gateways.http.response.ProdutoResponse;
-import com.good.food.usecase.produto.BuscarProdutoPorCategoria;
-import com.good.food.usecase.produto.CadastrarProduto;
-import com.good.food.usecase.produto.EditarProduto;
-import com.good.food.usecase.produto.RemoverProduto;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
